requirements.txt
numpy>=1.26.4,<3.0
Pillow>=10.4.0
opencv-python>=4.10.0
scikit-learn>=1.5.1
torch>=2.4.1
torchvision>=0.19.1
transformers>=4.46.0
safetensors>=0.4.5
ftfy>=6.3.1
regex>=2024.9.11
tqdm>=4.66.5
ultralytics>=8.0.0
diffusers>=0.30.0
accelerate>=0.34.0
git+https://github.com/openai/CLIP.git
basicsr>=1.4.2
realesrgan>=0.3.0
timm>=0.9.0
